+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 
 * 1.16
+    * New: Completely new default HTML emails.
     * New: "Waiting for payment" transaction status for Cash on Delivery, Bank Transfer and Cheque payments.
     * New: More actions for emails - now available all actions for both customers and admins.
     * New: Default email templates are now translatable.